Introduce BackupFileUtil
