+++ src/main/java/org/jabref/gui/dialogs/BackupUIManager.java
@@ -7,7 +7,7 @@ import org.jabref.gui.util.DefaultTaskExecutor;
 import org.jabref.logic.autosaveandbackup.BackupManager;
 import org.jabref.logic.l10n.Localization;
 import org.jabref.logic.util.BackupFileType;
-import org.jabref.logic.util.io.FileUtil;
+import org.jabref.logic.util.io.BackupFileUtil;
 
 /**
  * Stores all user dialogs related to {@link BackupManager}.
@@ -22,7 +22,7 @@ public class BackupUIManager {
                 .append(Localization.lang("A backup file for '%0' was found at '%1'.",
                         originalPath.getFileName().toString(),
                         // We need to determine the path "manually" as the path does not get passed through when a diff is detected.
-                        FileUtil.getPathOfLatestExisingBackupFile(originalPath, BackupFileType.BACKUP).map(Path::toString).orElse(Localization.lang("File not found"))))
+                        BackupFileUtil.getPathOfLatestExisingBackupFile(originalPath, BackupFileType.BACKUP).map(Path::toString).orElse(Localization.lang("File not found"))))
                 .append("\n")
                 .append(Localization.lang("This could indicate that JabRef did not shut down cleanly last time the file was used."))
                 .append("\n\n")
