+++ src/main/java/org/jabref/logic/util/io/FileUtil.java
@@ -3,19 +3,15 @@ package org.jabref.logic.util.io;
 import java.io.File;
 import java.io.IOException;
 import java.io.UncheckedIOException;
-import java.nio.ByteBuffer;
 import java.nio.file.CopyOption;
 import java.nio.file.FileSystems;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.StandardCopyOption;
 import java.nio.file.StandardOpenOption;
-import java.time.ZonedDateTime;
-import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
-import java.util.HexFormat;
 import java.util.List;
 import java.util.Locale;
 import java.util.Objects;
@@ -27,19 +23,20 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import org.jabref.logic.citationkeypattern.BracketedPattern;
-import org.jabref.logic.util.BackupFileType;
-import org.jabref.logic.util.BuildInfo;
 import org.jabref.model.database.BibDatabase;
 import org.jabref.model.database.BibDatabaseContext;
 import org.jabref.model.entry.BibEntry;
 import org.jabref.model.util.FileHelper;
 import org.jabref.model.util.OptionalUtil;
 
-import net.harawata.appdirs.AppDirsFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
+ * The idea of this class is to add general functionality that could possibly even in the
+ * <a href="https://en.wikipedia.org/wiki/Non-blocking_I/O_(Java)">Java NIO package</a>,
+ * such as getting/adding file extension etc.
+ *
  * This class is the "successor" of {@link FileHelper}. In case you miss something here,
  * please look at {@link FileHelper} and migrate the functionality to here.
  */
@@ -47,6 +44,7 @@ public class FileUtil {
 
     public static final boolean IS_POSIX_COMPLIANT = FileSystems.getDefault().supportedFileAttributeViews().contains("posix");
     public static final int MAXIMUM_FILE_NAME_LENGTH = 255;
+
     private static final Logger LOGGER = LoggerFactory.getLogger(FileUtil.class);
 
     private FileUtil() {
@@ -130,103 +128,6 @@ public class FileUtil {
                                       && !part.equals(fileName) && part.contains(File.separator))
                               .findFirst()
                               .map(part -> part.substring(0, part.lastIndexOf(File.separator)));
-    }
-
-    public static Path getAppDataBackupDir() {
-        Path directory = Path.of(AppDirsFactory.getInstance().getUserDataDir(
-                                     "jabref",
-                                     new BuildInfo().version.toString(),
-                                     "org.jabref"))
-                             .resolve("backups");
-        return directory;
-    }
-
-    /**
-     * Determines the path of the backup file (using the given extension)
-     *
-     * <p>
-     *     As default, a directory inside the user temporary dir is used.<br>
-     *     In case a AUTOSAVE backup is requested, a timestamp is added
-     * </p>
-     * <p>
-     *     <em>SIDE EFFECT</em>: Creates the directory.
-     *     In case that fails, the return path of the .bak file is set to be next to the .bib file
-     * </p>
-     * <p>
-     *     Note that this backup is different from the <code>.sav</code> file generated by {@link org.jabref.logic.autosaveandbackup.BackupManager}
-     *     (and configured in the preferences as "make backups")
-     * </p>
-     */
-    public static Path getPathForNewBackupFileAndCreateDirectory(Path targetFile, BackupFileType fileType) {
-        String extension = "." + fileType.getExtensions().get(0);
-        String timeSuffix = ZonedDateTime.now().format(DateTimeFormatter.ofPattern("yyyy-MM-dd--HH.mm.ss"));
-
-        // We choose the data directory, because a ".bak" file should survive cache cleanups
-        Path directory = getAppDataBackupDir();
-        try {
-            Files.createDirectories(directory);
-        } catch (IOException e) {
-            Path result = FileUtil.addExtension(targetFile, extension);
-            LOGGER.warn("Could not create bib writing directory {}, using {} as file", directory, result, e);
-            return result;
-        }
-        String baseFileName = getUniqueFilePrefix(targetFile) + "--" + targetFile.getFileName() + "--" + timeSuffix;
-        Path fileName = addExtension(Path.of(baseFileName), extension);
-        return directory.resolve(fileName);
-    }
-
-    /**
-     * Finds the latest backup (.sav). If it does not exist, an empty optional is returned
-     */
-    public static Optional<Path> getPathOfLatestExisingBackupFile(Path targetFile, BackupFileType fileType) {
-        // The code is similar to "getPathForNewBackupFileAndCreateDirectory"
-
-        String extension = "." + fileType.getExtensions().get(0);
-
-        Path directory = getAppDataBackupDir();
-        if (Files.notExists(directory)) {
-            // In case there is no app directory, we search in the directory of the bib file
-            Path result = FileUtil.addExtension(targetFile, extension);
-            if (Files.exists(result)) {
-                return Optional.of(result);
-            } else {
-                return Optional.empty();
-            }
-        }
-
-        // Search the directory for the latest file
-        final String prefix = FileUtil.getUniqueFilePrefix(targetFile) + "--" + targetFile.getFileName();
-        Optional<Path> mostRecentFile;
-        try {
-            mostRecentFile = Files.list(directory)
-                                         // just list the .sav belonging to the given targetFile
-                                         .filter(p -> p.getFileName().toString().startsWith(prefix))
-                                         .sorted()
-                                         .reduce((first, second) -> second);
-        } catch (IOException e) {
-            LOGGER.error("Could not determine most recent file", e);
-            return Optional.empty();
-        }
-        return mostRecentFile;
-    }
-
-    /**
-     * <p>
-     * Determines a unique file prefix.
-     * </p>
-     * <p>
-     *     When creating a backup file, the backup file should belong to the original file.
-     *     Just adding ".bak" suffix to the filename, does not work in all cases:
-     *     It may be possible that the user has opened "paper.bib" twice.
-     *     Thus, we need to create a unique prefix to distinguish these files.
-     * </p>
-     */
-    public static String getUniqueFilePrefix(Path targetFile) {
-        // Idea: use the hash code and convert it to hex
-        // Thereby, use positive values only and use length 4
-        int positiveCode = Math.abs(targetFile.hashCode());
-        byte[] array = ByteBuffer.allocate(4).putInt(positiveCode).array();
-        return HexFormat.of().formatHex(array);
     }
 
     /**
