+++ src/test/java/org/jabref/logic/autosaveandbackup/BackupManagerTest.java
@@ -5,7 +5,7 @@ import java.nio.file.Path;
 import java.nio.file.StandardCopyOption;
 
 import org.jabref.logic.util.BackupFileType;
-import org.jabref.logic.util.io.FileUtil;
+import org.jabref.logic.util.io.BackupFileUtil;
 
 import org.junit.jupiter.api.Test;
 
@@ -20,7 +20,7 @@ public class BackupManagerTest {
         Path bibPath = Path.of("tmp", "test.bib");
         Path bakPath = BackupManager.getBackupPathForNewBackup(bibPath);
 
-        assertEquals(FileUtil.getAppDataBackupDir(), bakPath.getParent());
+        assertEquals(BackupFileUtil.getAppDataBackupDir(), bakPath.getParent());
         String start = bakPath.getFileName().toString().substring(0, 20); // Timestamp will differ
         assertEquals("27182d3c--test.bib--", start);
     }
@@ -35,7 +35,7 @@ public class BackupManagerTest {
     public void backupFileIsEqual() throws Exception {
         // Prepare test: Create backup file on "right" path
         Path source = Path.of(BackupManagerTest.class.getResource("no-changes.bib.bak").toURI());
-        Path target = FileUtil.getPathForNewBackupFileAndCreateDirectory(Path.of(BackupManagerTest.class.getResource("no-changes.bib").toURI()), BackupFileType.BACKUP);
+        Path target = BackupFileUtil.getPathForNewBackupFileAndCreateDirectory(Path.of(BackupManagerTest.class.getResource("no-changes.bib").toURI()), BackupFileType.BACKUP);
         Files.copy(source, target, StandardCopyOption.REPLACE_EXISTING);
 
         Path originalFile = Path.of(BackupManagerTest.class.getResource("no-changes.bib").toURI());
@@ -46,7 +46,7 @@ public class BackupManagerTest {
     public void backupFileDiffers() throws Exception {
         // Prepare test: Create backup file on "right" path
         Path source = Path.of(BackupManagerTest.class.getResource("changes.bib.bak").toURI());
-        Path target = FileUtil.getPathForNewBackupFileAndCreateDirectory(Path.of(BackupManagerTest.class.getResource("changes.bib").toURI()), BackupFileType.BACKUP);
+        Path target = BackupFileUtil.getPathForNewBackupFileAndCreateDirectory(Path.of(BackupManagerTest.class.getResource("changes.bib").toURI()), BackupFileType.BACKUP);
         Files.copy(source, target, StandardCopyOption.REPLACE_EXISTING);
 
         Path originalFile = Path.of(BackupManagerTest.class.getResource("changes.bib").toURI());
@@ -59,15 +59,15 @@ public class BackupManagerTest {
 
         // most recent file does not have any changes
         Path source = Path.of(BackupManagerTest.class.getResource("no-changes.bib.bak").toURI());
-        Path target = FileUtil.getPathForNewBackupFileAndCreateDirectory(Path.of(BackupManagerTest.class.getResource("no-changes.bib").toURI()), BackupFileType.BACKUP);
+        Path target = BackupFileUtil.getPathForNewBackupFileAndCreateDirectory(Path.of(BackupManagerTest.class.getResource("no-changes.bib").toURI()), BackupFileType.BACKUP);
         Files.copy(source, target, StandardCopyOption.REPLACE_EXISTING);
 
         // create "older" .bak files containing changes
         for (int i = 0; i < 10; i++) {
             source = Path.of(BackupManagerTest.class.getResource("changes.bib").toURI());
-            Path directory = FileUtil.getAppDataBackupDir();
+            Path directory = BackupFileUtil.getAppDataBackupDir();
             String timeSuffix = "2020-02-03--00.00.0" + Integer.toString(i);
-            String fileName = FileUtil.getUniqueFilePrefix(Path.of(BackupManagerTest.class.getResource("no-changes.bib").toURI())) + "--no-changes.bib--" + timeSuffix + ".bak";
+            String fileName = BackupFileUtil.getUniqueFilePrefix(Path.of(BackupManagerTest.class.getResource("no-changes.bib").toURI())) + "--no-changes.bib--" + timeSuffix + ".bak";
             target = directory.resolve(fileName);
             Files.copy(source, target, StandardCopyOption.REPLACE_EXISTING);
         }
